Set test nonlinear system model as a pytest fixture
- Move `test_nonlinear_system_model` from
  `test_nonlinear_dd_mpc_integration.py` to `tests/conftest.py` and
  define it as a pytest fixture.
- Update `test_nonlinear_dd_mpc_integration` to use
  `test_nonlinear_system_model` as a fixture.

This prevents pytest from considering `test_nonlinear_system_model`
as a test function when running unmarked tests (unit tests), causing
the error "Failed: Expected None, but test returned..."
